api: change ngl_easing_*() nb_args argument type to size_t

+++ libnopegl/src/node_animkeyframe.c
@@ -490,7 +490,7 @@ static char *animkeyframe_info_str(const struct ngl_node *node)
     return ret;
 }
 
-int ngl_easing_evaluate(const char *name, const double *args, int nb_args,
+int ngl_easing_evaluate(const char *name, const double *args, size_t nb_args,
                         const double *offsets, double t, double *v)
 {
     int easing_id;
@@ -517,7 +517,7 @@ int ngl_easing_evaluate(const char *name, const double *args, int nb_args,
     return 0;
 }
 
-int ngl_easing_derivate(const char *name, const double *args, int nb_args,
+int ngl_easing_derivate(const char *name, const double *args, size_t nb_args,
                         const double *offsets, double t, double *v)
 {
     int easing_id;
@@ -545,7 +545,7 @@ int ngl_easing_derivate(const char *name, const double *args, int nb_args,
     return 0;
 }
 
-int ngl_easing_solve(const char *name, const double *args, int nb_args,
+int ngl_easing_solve(const char *name, const double *args, size_t nb_args,
                      const double *offsets, double v, double *t)
 {
     int easing_id;
