+++ pynopegl/_pynopegl.pyx
@@ -188,11 +188,11 @@ cdef extern from "nopegl.h":
     void ngl_livectls_freep(ngl_livectl **livectlsp)
     void ngl_freep(ngl_ctx **ss)
 
-    int ngl_easing_evaluate(const char *name, const double *args, int nb_args,
+    int ngl_easing_evaluate(const char *name, const double *args, size_t nb_args,
                             const double *offsets, double t, double *v)
-    int ngl_easing_derivate(const char *name, const double *args, int nb_args,
+    int ngl_easing_derivate(const char *name, const double *args, size_t nb_args,
                             const double *offsets, double t, double *v)
-    int ngl_easing_solve(const char *name, const double *args, int nb_args,
+    int ngl_easing_solve(const char *name, const double *args, size_t nb_args,
                          const double *offsets, double v, double *t)
 
     int ngl_gl_wrap_framebuffer(ngl_ctx *s, uint32_t framebuffer)
@@ -400,7 +400,7 @@ ANIM_EVALUATE, ANIM_DERIVATE, ANIM_SOLVE = range(3)
 def animate(const char *name, double src, args, offsets, mode):
     cdef double c_args[2]
     cdef double *c_args_param = NULL
-    cdef int nb_args = 0
+    cdef size_t nb_args = 0
     if args is not None:
         nb_args = len(args)
         if nb_args > 2:
